chore: make clippy happy (#12)

+++ run_loop/src/platform/win32/adapter.rs
@@ -69,7 +69,7 @@ unsafe extern "system" fn wnd_proc(
         let bridge = &*(ptr as *const EventBridge);
         let handler = &*(bridge.handler);
         let res = handler.wnd_proc(h_wnd, msg, w_param, l_param);
-        if msg == WM_NCDESTROY as u32 {
+        if msg == WM_NCDESTROY {
             // make sure bridge is dropped
             let _ = Box::<EventBridge>::from_raw(ptr as *mut EventBridge);
         }
